Minor discrepancies fixed
In the description for assignment 1 - "Default crop size is 150 but allow it to be variable" is specified, but the code has default patch size to be 100. Fixed this to be consistent. 

In the description for assignment 2, the batch size is set to be 1 - meaning the size of the output should be 1, 200, 200, 200. But, the solution has batch size of 3, and describes the output to be either 2, 200, 200, 200 or 3, 200, 200, 200. This could be confusing - and hence updated the code to specify batch size to be 1 - so that everything is consistent.